Move swiglu to a util file + add optional residual flag
