+++ aiter/ops/triton/_triton_kernels/moe/moe_op_gemm_a8w8.py
@@ -6,6 +6,7 @@ import triton
 import triton.language as tl
 from aiter.ops.triton.utils._triton.pid_preprocessing import pid_grid
 from aiter.ops.triton._triton_kernels.moe.quant_moe import _compute_static_fp8_quant
+from aiter.ops.triton._triton_kernels.moe.activations import _swiglu
 
 
 def matmul_launch_metadata(grid, kernel, args):
@@ -105,27 +106,6 @@ def unswizzle_mx_scale_cdna4(
 
 
 @triton.jit
-def clip(x, limit, clip_lower: tl.constexpr):
-    res = tl.minimum(x, limit)
-    if clip_lower:
-        res = tl.maximum(-limit, res)
-    return res
-
-
-@triton.jit
-def _swiglu(input, alpha, limit):
-    gelu, linear = tl.split(tl.reshape(input, (input.shape[0], input.shape[1] // 2, 2)))
-    gelu = gelu.to(tl.float32)
-    if limit is not None:
-        gelu = clip(gelu, limit, clip_lower=False)
-    linear = linear.to(tl.float32)
-    if limit is not None:
-        linear = clip(linear, limit, clip_lower=True)
-    s = gelu / (1 + tl.exp2(-1.44269504089 * alpha * gelu))
-    return tl.fma(s, linear, s)  # (s * (linear + 1))
-
-
-@triton.jit
 def _reduce_grouped(
     X,
     stride_xb: tl.uint64,
@@ -145,6 +125,7 @@ def _reduce_grouped(
     K: tl.constexpr,
     BLOCK_N: tl.constexpr,
     EVEN_N: tl.constexpr,
+    ADD_RESIDUAL: tl.constexpr,
 ):
     pid_t = tl.program_id(1)
     pid_n = tl.program_id(0)
@@ -178,7 +159,7 @@ def _reduce_grouped(
 
         # apply nonlinearity to split-k output
         if APPLY_SWIGLU:
-            curr = _swiglu(curr[None, :], alpha, limit)
+            curr = _swiglu(curr[None, :], alpha, limit, ADD_RESIDUAL=ADD_RESIDUAL)
         curr = tl.reshape(curr, [curr.shape[-1]])
         # update final accumulator
         acc += curr
@@ -236,6 +217,7 @@ def _moe_gemm_a8w8(
     alpha,
     limit,
     ACTIVATION_REDUCTION_N: tl.constexpr,
+    ADD_RESIDUAL: tl.constexpr,
     # MoE config
     N_EXPTS_ACT: tl.constexpr,
     # optimization config
@@ -491,7 +473,7 @@ def _moe_gemm_a8w8(
             bias = tl.full([BLOCK_N], 0, dtype=tl.float32)
         acc = acc + bias[None, :]
     if APPLY_SWIGLU and SPLIT_K == 1:
-        out = _swiglu(acc, alpha, limit)
+        out = _swiglu(acc, alpha, limit, ADD_RESIDUAL=ADD_RESIDUAL)
         tl.static_assert(
             out.shape[1] == OUT_BLOCK_N,
             f"Activation fn out.shape[1] ({out.shape[1]}) doesn't match computed OUT_BLOCK_N ({OUT_BLOCK_N})",
