+++ aiter/ops/triton/moe/moe_op_gemm_a8w4.py
@@ -140,6 +140,7 @@ def reduce_grouped(
     limit=1.0,
     reduction_n=1,
     out_dtype: bool = None,
+    add_residual: bool = True,
 ):
     """
     In-place grouped row reduction.
@@ -200,6 +201,7 @@ def reduce_grouped(
         BLOCK_N=BLOCK_N,
         EVEN_N=(x.shape[-1] % BLOCK_N == 0),
         K=K,  #
+        ADD_RESIDUAL=add_residual,
         num_warps=2,  #
     )
     return out
@@ -227,6 +229,7 @@ def moe_gemm_a8w4(
     apply_swiglu=False,
     alpha=1.0,
     limit=1.0,
+    add_residual=True,
     unpadded_N=None,
     unpadded_K=None,
 ):
@@ -332,6 +335,7 @@ def moe_gemm_a8w4(
         alpha,
         limit,
         reduction_n_matmul,
+        add_residual,
         routing_data.n_expts_act,
         config["block_m"],
         config["block_n"],
@@ -365,6 +369,7 @@ def moe_gemm_a8w4(
         limit,
         reduction_n_reduction,
         out_dtype=out_dtype,
+        add_residual=add_residual,
     )
     return y_final
 
@@ -374,7 +379,7 @@ def moe_gemm_a8w4(
 # -----------------------------------------------------------------------------
 
 
-def swiglu_torch(a, alpha, limit):
+def swiglu_torch(a, alpha, limit, add_residual=True):
     a_gelu = a[..., ::2]
     if limit is not None:
         a_gelu = a_gelu.clamp(max=limit)
@@ -383,7 +388,10 @@ def swiglu_torch(a, alpha, limit):
         a_linear = a_linear.clamp(min=-limit, max=limit)
 
     out_gelu = a_gelu * torch.sigmoid(alpha * a_gelu)
-    out = out_gelu * (a_linear + 1)
+    if add_residual:
+        out = out_gelu * (a_linear + 1)
+    else:
+        out = out_gelu * a_linear
     return out
 
 
@@ -398,6 +406,7 @@ def moe_gemm_torch(
     apply_swiglu=False,
     alpha=1.0,
     limit=1.0,
+    add_residual=True,
 ):
     assert x.dtype.itemsize > 1
     assert w.dtype.itemsize > 1
@@ -427,7 +436,7 @@ def moe_gemm_torch(
         if bias is not None:
             out += bias[i, :]
         if apply_swiglu:
-            out = swiglu_torch(out, alpha, limit)
+            out = swiglu_torch(out, alpha, limit, add_residual)
         if gammas is not None:
             out *= gammas[lo:hi, None]
         y[lo:hi, :] = out
