+++ aiter/ops/triton/moe/moe_op_gemm_a8w8.py
@@ -130,6 +130,7 @@ def reduce_grouped(
     limit=1.0,
     reduction_n=1,
     out_dtype: bool = None,
+    add_residual: bool = True,
 ):
     """
     In-place grouped row reduction.
@@ -190,6 +191,7 @@ def reduce_grouped(
         BLOCK_N=BLOCK_N,
         EVEN_N=(x.shape[-1] % BLOCK_N == 0),
         K=K,  #
+        ADD_RESIDUAL=add_residual,
         num_warps=2,  #
     )
     return out
@@ -218,6 +220,7 @@ def moe_gemm_a8w8(
     apply_swiglu=False,
     alpha=1.0,
     limit=1.0,
+    add_residual=True,
     unpadded_N=None,
     unpadded_K=None,
 ):
@@ -335,6 +338,7 @@ def moe_gemm_a8w8(
         alpha,
         limit,
         reduction_n_matmul,
+        add_residual,
         routing_data.n_expts_act,
         config["block_m"],
         config["block_n"],
@@ -368,6 +372,7 @@ def moe_gemm_a8w8(
         limit,
         reduction_n_reduction,
         out_dtype=out_dtype,
+        add_residual=add_residual,
     )
     return y_final
 
@@ -377,7 +382,7 @@ def moe_gemm_a8w8(
 # -----------------------------------------------------------------------------
 
 
-def swiglu_torch(a, alpha, limit):
+def swiglu_torch(a, alpha, limit, add_residual=True):
     a_gelu = a[..., ::2]
     if limit is not None:
         a_gelu = a_gelu.clamp(max=limit)
@@ -386,7 +391,10 @@ def swiglu_torch(a, alpha, limit):
         a_linear = a_linear.clamp(min=-limit, max=limit)
 
     out_gelu = a_gelu * torch.sigmoid(alpha * a_gelu)
-    out = out_gelu * (a_linear + 1)
+    if add_residual:
+        out = out_gelu * (a_linear + 1)
+    else:
+        out = out_gelu * a_linear
     return out
 
 
@@ -401,6 +409,7 @@ def moe_gemm_torch(
     apply_swiglu=False,
     alpha=1.0,
     limit=1.0,
+    add_residual=True,
 ):
     assert x.dtype.itemsize > 1
     assert w.dtype.itemsize > 1
@@ -430,7 +439,7 @@ def moe_gemm_torch(
         if bias is not None:
             out += bias[i, :]
         if apply_swiglu:
-            out = swiglu_torch(out, alpha, limit)
+            out = swiglu_torch(out, alpha, limit, add_residual)
         if gammas is not None:
             out *= gammas[lo:hi, None]
         y[lo:hi, :] = out
