+++ aiter/ops/triton/moe/moe_op_gemm_a8w8_blockscale.py
@@ -123,6 +123,7 @@ def reduce_grouped(
     limit=None,
     reduction_n=1,
     out_dtype: bool = None,
+    add_residual: bool = True,
 ):
     """
     In-place grouped row reduction.
@@ -183,6 +184,7 @@ def reduce_grouped(
         BLOCK_N=BLOCK_N,
         EVEN_N=(x.shape[-1] % BLOCK_N == 0),
         K=K,  #
+        ADD_RESIDUAL=add_residual,
         num_warps=2,  #
     )
     return out
@@ -210,6 +212,7 @@ def moe_gemm_a8w8_blockscale(
     apply_swiglu=False,
     alpha=1.0,
     limit=None,
+    add_residual=True,
     unpadded_N=None,
     unpadded_K=None,
     per_row_x_scale=False,
@@ -326,6 +329,7 @@ def moe_gemm_a8w8_blockscale(
         alpha,
         limit,
         reduction_n_matmul,
+        add_residual,
         routing_data.n_expts_act,
         config["block_m"],
         config["block_n"],
@@ -362,6 +366,7 @@ def moe_gemm_a8w8_blockscale(
         limit,
         reduction_n_reduction,
         out_dtype=out_dtype,
+        add_residual=add_residual,
     )
     return y_final
 
@@ -371,7 +376,7 @@ def moe_gemm_a8w8_blockscale(
 # -----------------------------------------------------------------------------
 
 
-def swiglu_torch(a, alpha, limit):
+def swiglu_torch(a, alpha, limit, add_residual=True):
     a_gelu = a[..., ::2]
     if limit is not None:
         a_gelu = a_gelu.clamp(max=limit)
@@ -380,7 +385,10 @@ def swiglu_torch(a, alpha, limit):
         a_linear = a_linear.clamp(min=-limit, max=limit)
 
     out_gelu = a_gelu * torch.sigmoid(alpha * a_gelu)
-    out = out_gelu * (a_linear + 1)
+    if add_residual:
+        out = out_gelu * (a_linear + 1)
+    else:
+        out = out_gelu * a_linear
     return out
 
 
@@ -395,6 +403,7 @@ def moe_gemm_torch(
     apply_swiglu=False,
     alpha=1.0,
     limit=None,
+    add_residual=True,
 ):
     assert x.dtype.itemsize > 1
     assert w.dtype.itemsize > 1
@@ -424,7 +433,7 @@ def moe_gemm_torch(
         if bias is not None:
             out += bias[i, :]
         if apply_swiglu:
-            out = swiglu_torch(out, alpha, limit)
+            out = swiglu_torch(out, alpha, limit, add_residual)
         if gammas is not None:
             out *= gammas[lo:hi, None]
         y[lo:hi, :] = out
